+++ model/geo_algorithms/src/distance/primitive_3d.rs
@@ -380,7 +380,8 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source =
+            include_str!("../collision/primitive_3d/cylindrical_and_conical_family.rs");
         let intersection_source =
             include_str!("../intersection/primitive_3d/cylindrical_and_conical_family.rs");
         let collision_cylindrical_surface_point_section = section(
@@ -469,7 +470,8 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source =
+            include_str!("../collision/primitive_3d/cylindrical_and_conical_family.rs");
         let collision_cylindrical_solid_point_section = section(
             collision_source,
             "pub fn cylindrical_solid3d_point3d_collides",
@@ -503,7 +505,7 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source = include_str!("../collision/primitive_3d/circular_family.rs");
         let intersection_source = include_str!("../intersection/primitive_3d/circular_family.rs");
         let collision_ellipse_point_section = section(
             collision_source,
@@ -550,7 +552,7 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source = include_str!("../collision/primitive_3d/circular_family.rs");
         let intersection_source = include_str!("../intersection/primitive_3d/circular_family.rs");
         let collision_arc_point_section = section(
             collision_source,
@@ -598,13 +600,12 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source = include_str!("../collision/primitive_3d/circular_family.rs");
         let intersection_source = include_str!("../intersection/primitive_3d/circular_family.rs");
-        let collision_circle_point_section = section(
-            collision_source,
-            "pub fn circle3d_point3d_collides",
-            "pub fn plane3d_point3d_collides",
-        );
+        let collision_circle_point_start = collision_source
+            .find("pub fn circle3d_point3d_collides")
+            .unwrap_or_else(|| panic!("missing start marker: pub fn circle3d_point3d_collides"));
+        let collision_circle_point_section = &collision_source[collision_circle_point_start..];
         let intersection_circle_point_section = section(
             intersection_source,
             "fn circle3d_point3d_intersection_raw",
@@ -652,7 +653,8 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source =
+            include_str!("../collision/primitive_3d/spherical_and_quadric_family.rs");
         let intersection_source =
             include_str!("../intersection/primitive_3d/spherical_and_quadric_family.rs");
         let collision_ellipsoidal_solid_point_section = section(
@@ -707,7 +709,8 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source =
+            include_str!("../collision/primitive_3d/spherical_and_quadric_family.rs");
         let intersection_source =
             include_str!("../intersection/primitive_3d/spherical_and_quadric_family.rs");
         let collision_ellipsoidal_surface_point_section = section(
@@ -795,13 +798,14 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source =
+            include_str!("../collision/primitive_3d/spherical_and_quadric_family.rs");
         let intersection_source =
             include_str!("../intersection/primitive_3d/spherical_and_quadric_family.rs");
         let collision_spherical_solid_point_section = section(
             collision_source,
             "pub fn spherical_solid3d_point3d_collides",
-            "pub fn cylindrical_solid3d_point3d_collides",
+            "pub fn spherical_solid3d_circle3d_collides",
         );
         let intersection_spherical_solid_point_section = section(
             intersection_source,
@@ -844,7 +848,8 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source =
+            include_str!("../collision/primitive_3d/spherical_and_quadric_family.rs");
         let intersection_source =
             include_str!("../intersection/primitive_3d/spherical_and_quadric_family.rs");
         let collision_torus_solid_point_section = section(
@@ -895,14 +900,17 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source =
+            include_str!("../collision/primitive_3d/spherical_and_quadric_family.rs");
         let intersection_source =
             include_str!("../intersection/primitive_3d/spherical_and_quadric_family.rs");
-        let collision_torus_surface_point_section = section(
-            collision_source,
-            "pub fn torus_surface3d_point3d_collides",
-            "pub fn triangle3d_point3d_collides",
-        );
+        let collision_torus_surface_point_start = collision_source
+            .find("pub fn torus_surface3d_point3d_collides")
+            .unwrap_or_else(|| {
+                panic!("missing start marker: pub fn torus_surface3d_point3d_collides")
+            });
+        let collision_torus_surface_point_section =
+            &collision_source[collision_torus_surface_point_start..];
         let intersection_torus_surface_point_section = section(
             intersection_source,
             "fn torus_surface3d_point3d_intersection_raw",
@@ -942,7 +950,7 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source = include_str!("../collision/primitive_3d/planar_and_mesh_family.rs");
         let intersection_source =
             include_str!("../intersection/primitive_3d/planar_and_mesh_family.rs");
         let collision_triangle_point_section = section(
@@ -983,13 +991,13 @@ mod tests {
             &tail[..end_index]
         }
 
-        let collision_source = include_str!("../collision/primitive_3d.rs");
+        let collision_source = include_str!("../collision/primitive_3d/planar_and_mesh_family.rs");
         let intersection_source =
             include_str!("../intersection/primitive_3d/planar_and_mesh_family.rs");
         let collision_triangle_mesh_point_section = section(
             collision_source,
             "pub fn triangle_mesh3d_point3d_collides",
-            "pub fn arc3d_point3d_collides",
+            "pub fn plane3d_point3d_collides",
         );
         let intersection_triangle_mesh_point_section = section(
             intersection_source,
